@@ -68,14 +68,10 @@ under the License.
   </distributionManagement>
 
   <properties>
-    <javaVersion>8</javaVersion>
-    <mavenVersion>3.9.6</mavenVersion>
-    <!-- Keep in sync with resolver used in maven above -->
-    <slf4jVersion>1.7.36</slf4jVersion>
-    <!-- Keep in sync with resolver used in maven above -->
-    <resolverVersion>1.9.20</resolverVersion>
+    <javaVersion>17</javaVersion>
+    <mavenVersion>4.0.0-beta-3</mavenVersion>
 
-    <!-- plugins version used in IT tests -->
+    <guiceVersion>6.0.0</guiceVersion>
     <mavenAntrunPluginVersion>${version.maven-antrun-plugin}</mavenAntrunPluginVersion>
     <mavenCompilerPluginVersion>${version.maven-compiler-plugin}</mavenCompilerPluginVersion>
     <mavenEnforcerPluginVersion>${version.maven-enforcer-plugin}</mavenEnforcerPluginVersion>
