@@ -18,25 +18,29 @@
  */
 package org.apache.maven.plugins.deploy;
 
-import org.apache.maven.execution.MavenSession;
-import org.apache.maven.plugin.AbstractMojo;
-import org.apache.maven.plugin.MojoExecutionException;
-import org.apache.maven.plugin.MojoFailureException;
-import org.apache.maven.plugins.annotations.Component;
-import org.apache.maven.plugins.annotations.Parameter;
-import org.apache.maven.rtinfo.RuntimeInformation;
-import org.eclipse.aether.RepositorySystem;
-import org.eclipse.aether.deployment.DeployRequest;
-import org.eclipse.aether.deployment.DeploymentException;
-import org.eclipse.aether.repository.RemoteRepository;
-import org.eclipse.aether.util.version.GenericVersionScheme;
-import org.eclipse.aether.version.InvalidVersionSpecificationException;
-import org.eclipse.aether.version.Version;
+import org.apache.maven.api.RemoteRepository;
+import org.apache.maven.api.Session;
+import org.apache.maven.api.Version;
+import org.apache.maven.api.di.Inject;
+import org.apache.maven.api.plugin.Log;
+import org.apache.maven.api.plugin.Mojo;
+import org.apache.maven.api.plugin.MojoException;
+import org.apache.maven.api.plugin.annotations.Parameter;
 
 /**
  * Abstract class for Deploy mojo's.
  */
-public abstract class AbstractDeployMojo extends AbstractMojo {
+public abstract class AbstractDeployMojo implements Mojo {
+    private static final String AFFECTED_MAVEN_PACKAGING = "maven-plugin";
+
+    private static final String FIXED_MAVEN_VERSION = "3.9.0";
+
+    @Inject
+    protected Log logger;
+
+    @Inject
+    protected Session session;
+
     /**
      * Flag whether Maven is currently in online/offline mode.
      */
@@ -52,45 +56,31 @@ public abstract class AbstractDeployMojo extends AbstractMojo {
     @Parameter(property = "retryFailedDeploymentCount", defaultValue = "1")
     private int retryFailedDeploymentCount;
 
-    @Component
-    private RuntimeInformation runtimeInformation;
-
-    @Parameter(defaultValue = "${session}", readonly = true, required = true)
-    protected MavenSession session;
-
-    @Component
-    protected RepositorySystem repositorySystem;
-
-    private static final String AFFECTED_MAVEN_PACKAGING = "maven-plugin";
-
-    private static final String FIXED_MAVEN_VERSION = "3.9.0";
-
     /* Setters and Getters */
 
-    void failIfOffline() throws MojoFailureException {
+    void failIfOffline() throws MojoException {
         if (offline) {
-            throw new MojoFailureException("Cannot deploy artifacts when Maven is in offline mode");
+            throw new MojoException("Cannot deploy artifacts when Maven is in offline mode");
         }
+    }
+
+    public int getRetryFailedDeploymentCount() {
+        return retryFailedDeploymentCount;
     }
 
     /**
      * If this plugin used in pre-3.9.0 Maven, the packaging {@code maven-plugin} will not deploy G level metadata.
      */
-    protected void warnIfAffectedPackagingAndMaven(final String packaging) {
+    protected void warnIfAffectedPackagingAndMaven(String packaging) {
         if (AFFECTED_MAVEN_PACKAGING.equals(packaging)) {
-            try {
-                GenericVersionScheme versionScheme = new GenericVersionScheme();
-                Version fixedMavenVersion = versionScheme.parseVersion(FIXED_MAVEN_VERSION);
-                Version currentMavenVersion = versionScheme.parseVersion(runtimeInformation.getMavenVersion());
-                if (fixedMavenVersion.compareTo(currentMavenVersion) > 0) {
-                    getLog().warn("");
-                    getLog().warn("You are about to deploy a maven-plugin using Maven " + currentMavenVersion + ".");
-                    getLog().warn("This plugin should be used ONLY with Maven 3.9.0 and newer, as MNG-7055");
-                    getLog().warn("is fixed in those versions of Maven only!");
-                    getLog().warn("");
-                }
-            } catch (InvalidVersionSpecificationException e) {
-                // skip it: Generic does not throw, only API contains this exception
+            Version fixedMavenVersion = session.parseVersion(FIXED_MAVEN_VERSION);
+            Version currentMavenVersion = session.getMavenVersion();
+            if (fixedMavenVersion.compareTo(currentMavenVersion) > 0) {
+                getLog().warn("");
+                getLog().warn("You are about to deploy a maven-plugin using Maven " + currentMavenVersion + ".");
+                getLog().warn("This plugin should be used ONLY with Maven 3.9.0 and newer, as MNG-7055");
+                getLog().warn("is fixed in those versions of Maven only!");
+                getLog().warn("");
             }
         }
     }
@@ -98,60 +88,15 @@ public abstract class AbstractDeployMojo extends AbstractMojo {
     /**
      * Creates resolver {@link RemoteRepository} equipped with needed whistles and bells.
      */
-    protected RemoteRepository getRemoteRepository(final String repositoryId, final String url) {
-        // TODO: RepositorySystem#newDeploymentRepository does this very same thing!
-        RemoteRepository result = new RemoteRepository.Builder(repositoryId, "default", url).build();
-
-        if (result.getAuthentication() == null || result.getProxy() == null) {
-            RemoteRepository.Builder builder = new RemoteRepository.Builder(result);
-
-            if (result.getAuthentication() == null) {
-                builder.setAuthentication(session.getRepositorySession()
-                        .getAuthenticationSelector()
-                        .getAuthentication(result));
-            }
-
-            if (result.getProxy() == null) {
-                builder.setProxy(
-                        session.getRepositorySession().getProxySelector().getProxy(result));
-            }
-
-            result = builder.build();
-        }
-
-        return result;
+    protected RemoteRepository createDeploymentArtifactRepository(String id, String url) {
+        return getSession().createRemoteRepository(id, url);
     }
 
-    // I'm not sure if retries will work with deploying on client level ...
-    // Most repository managers block a duplicate artifacts.
+    protected Session getSession() {
+        return session;
+    }
 
-    // Eg, when we have an artifact list, even simple pom and jar in one request with released version,
-    // next try can fail due to duplicate.
-
-    protected void deploy(DeployRequest deployRequest) throws MojoExecutionException {
-        int retryFailedDeploymentCounter = Math.max(1, Math.min(10, retryFailedDeploymentCount));
-        DeploymentException exception = null;
-        for (int count = 0; count < retryFailedDeploymentCounter; count++) {
-            try {
-                if (count > 0) {
-                    getLog().info("Retrying deployment attempt " + (count + 1) + " of " + retryFailedDeploymentCounter);
-                }
-
-                repositorySystem.deploy(session.getRepositorySession(), deployRequest);
-                exception = null;
-                break;
-            } catch (DeploymentException e) {
-                if (count + 1 < retryFailedDeploymentCounter) {
-                    getLog().warn("Encountered issue during deployment: " + e.getLocalizedMessage());
-                    getLog().debug(e);
-                }
-                if (exception == null) {
-                    exception = e;
-                }
-            }
-        }
-        if (exception != null) {
-            throw new MojoExecutionException(exception.getMessage(), exception);
-        }
+    protected Log getLog() {
+        return logger;
     }
 }
